lead tags added

+++ database/migrations/0000013_sales_management_create_lead_notes_table.php
@@ -13,16 +13,16 @@ return new class extends Migration {
      */
     public function up()
     {
-        Schema::create(config('sales-management.tablePrefix').'lead_notes', function (Blueprint $table) {
+        Schema::create(config('sales-management.tablePrefix') . 'lead_notes', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger("lead_id")->index();
-            $table->integer('created_by')->nullable();
+            $table->unsignedBigInteger('created_by')->nullable();
             $table->text("note")->nullable();
             $table->timestamps();
 
             $table->foreign('lead_id')
                 ->references('id')
-                ->on(config('sales-management.tablePrefix').'leads')
+                ->on(config('sales-management.tablePrefix') . 'leads')
                 ->cascadeOnDelete();
         });
     }
@@ -34,6 +34,6 @@ return new class extends Migration {
      */
     public function down()
     {
-        Schema::drop(config('sales-management.tablePrefix').'lead_notes');
+        Schema::drop(config('sales-management.tablePrefix') . 'lead_notes');
     }
 };
